+++ spec/examples/record/properties_spec.rb
@@ -7,6 +7,7 @@ describe Cequel::Record::Properties do
     model :Post do
       key :permalink, :text
       column :title, :text
+      column :status, :enum, values: { open: 1, closed: 2 }
       list :tags, :text
       set :categories, :text
       map :shares, :text, :int
@@ -28,6 +29,17 @@ describe Cequel::Record::Properties do
 
     it 'should have nil key if unset' do
       expect(Post.new.permalink).to be_nil
+    end
+
+    it 'should have enums' do
+      expect(Post.new.status).to be_nil
+      expect(Post.status).to eql({ open: 1, closed: 2 })
+      expect(Post.new { |post| post.status = :open }.status).to eq(:open)
+      expect(Post.new { |post| post.status = :open }).to be_open
+      expect(Post.new { |post| post.status = :closed }.status).to eq(:closed)
+      expect(Post.new { |post| post.status = :closed }).to be_closed
+      expect(Post.new { |post| post.status = :closed }.attributes['status']).to eq(2)
+      expect(Post.new { |post| post.status = :unknown }.status).to eq(nil)
     end
 
     it 'should provide accessor for data column' do
@@ -197,7 +209,7 @@ describe Cequel::Record::Properties do
     end
 
     it 'should run default proc on keys' do
-      expect(Post.new.subid).to eq("subid #{1+1}") 
+      expect(Post.new.subid).to eq("subid #{1+1}")
     end
 
     it 'should run default proc' do